Reduced the number of sort wrapper applications even further: When producing assertions, an analysis determines where the snapshot (or rather, the snapshot generator) passed to produce will eventually be used, that is, which actual sort the snapshot is supposed to have. E.g., when producing an access predicate `acc(x.f)`, where `f: Int`, the snapshot should be of sort `Int`. When producing an access predicate to a predicate, e.g., `acc(P(x))`, the predicate body is analysed accordingly so that, if `predicate P(x) {acc(x.f)}`, the snapshot given to the chunk corresponding to `P(x)` will be of sort `Int`.
It is (currently) not possible to completely remove sort wrappers that go from `Snap` to a concrete sort, e.g., `Int`, because of predicates with a body such as `predicate Q(x, b) {b ? acc(x.f) : acc(x.g)}`, where `f: Int` and `g: Bool`. See test `all/predicates/different_field_types.sil` for an example.
The problem is that the snapshot given to `Q` can neither be `Int` nor `Bool`, so for now, it will be `Snap`, which is ok as long as wrapping it to `Int` and to `Bool` is made on disjoint paths through the program.

It should be possible to reduce the number of sort wrapper applications even further by using a similar analysis before declaring functions coming from the SIL program on the prover level. Currently, the snapshot of the function is fixed to sort `Snap`, and thus, each computed snapshot is wrapped to `Snap` before it is used in a function application term.

Another source of wrapper applications are snapshot pairs (`combine`), since that requires the paired snapshots to be