Reworked the fvf axioms. In essence, the change is that axioms previously of the form:
forall i:T :: P(e(i)) ==> fvf(e(i)) == T(e(i))
(in which T could be another fvf or other snapshot, and in which e(i) is the receiver expression occurring in the term trying to be evaluated/exhaled)

are now generated in the form:

forall r:Ref :: P(r) ==> fvf(r) == T(r)

with correspondingly simpler triggers. This means that the axioms match in situations in which a differently-structured expression is "looked-up" in a field-value function, which seems necessary for interesting examples to have a chance of making use of these axioms.

In the code, fresh quantified variables are generated where the fvfDefs are generated. This largely eliminates the need for passing the quantified variable in the original receiver expression (the "i" above), but that still seems to be used for permission checking - I'm not sure how, so have left it there.
Because such a fresh variable should not be used in the case in which the fvfDefs are not to be used in a quantifier (where the original receiver term e must be ground already), a flag "needsQuantifying" is added to the signature of withValue (this is clear in the context the method is called).

Unfortunately, the extra triggering seems to slow down some tests a lot. In particular, performance.sil seems to be extremely slow now. I'm a bit surprised by that, since that test case doesn't seem to require fvfs (it's just about permission accounting).

Anyway, obviously if we decide that we don't like the change, we can back it out. But it allows me to make progress with a competition example which was stuck very early on with incompletenesses.