Improved handling and reporting of Java's `Exception`s and `Error`s: exceptions are turned into Viper verification failures, errors are still thrown. In addition, for a `NoClassDefFoundError`, a message with instructions of how to possibly overcome the problem is shown.
Also replaced all uses of Scala's `???` by Scala's `sys.error`: the former raised an `Error` whereas the latter raises a `RuntimeException`.

The convention should be that any unexpected/unhandled/erroneous situation for which a custom failure message is (in essentially all situations) better than a stack trace is reported as the former. There should be very few cases in which a custom failure message is not clearly better than a stack trace, e.g. for OutOfMemory errors and for aforementioned `NoClassDefFoundError`s.