Set StdIOReporter as default Reporter for Silcion. Keep NoopReporter for Silicon tests.
We duplicate all reports into logs. By-default, the logs should be written to a (temporary, automatically selected) log file. The default reports, however, must be (carefully) printed to STDOUT. StdIOReporter should take care of that. Therefore, the set of logs is a superset of the set of reports (modulo copyrights, tool sygnatures, etc.)