Put old trace behaviour under a new flag.
The old execution log behaviour (trace written to the executionTreeData.js file)
is still available when the "writeTraceFile" flag is set. The "ideModeAdvanced"
is now used to enable the new execution trace format and to send the trace via
ViperServer. The two flags are independent, i.e. it is possible to write the
trace in old format to the executionTreedata.js file, and at the same time to
send it in the new format through ViperServer.