Add emission of BlockFailureMessages, clean up a bit
There were issues with the --numErrorsToReport option.
PathProcessedMessages may have emitted false results since the result of
the last explored block did not match the expected path result. It was
success if the last explored block succeeded, while it was expected to
be Failure if any block on the path failed. Now the message should
merely signify that the path has finished.